+++ rust-bao/pkg/bao_wasm_bg.wasm.d.ts
@@ -1,0 +1,13 @@
+/* tslint:disable */
+/* eslint-disable */
+export const memory: WebAssembly.Memory;
+export const chunk_cv: (a: number, b: bigint, c: number) => void;
+export const get_input_ptr: () => number;
+export const get_input_size: () => number;
+export const get_output_ptr: () => number;
+export const parent_cv: (a: number) => void;
+export const batch_parent_cvs: (a: number, b: number) => void;
+export const get_output_size: () => number;
+export const batch_chunk_cvs: (a: number, b: bigint) => void;
+export const __wbindgen_externrefs: WebAssembly.Table;
+export const __wbindgen_start: () => void;